Refactoring appearance form
