+++ code/recipe-book-app/src/components/index.js
@@ -1,5 +1,12 @@
 import RecipeForm from "./RecipeForm";
 import FavoriteButton from "./FavoriteButton";
+import {
+	AutoCompleteWrapper,
+	TextfieldWrapper,
+	SubmitButtonWrapper,
+	ToggleSwitchWrapper,
+	ResetButtonWrapper,
+} from "./FormUI";
 import Header from "./Header";
 import ImageGallery from "./ImageGallery";
 import Navigation from "./Navigation";
@@ -9,6 +16,11 @@ import ViewRecipe from "./ViewRecipe";
 export {
 	RecipeForm,
 	FavoriteButton,
+	AutoCompleteWrapper,
+	TextfieldWrapper,
+	SubmitButtonWrapper,
+	ToggleSwitchWrapper,
+	ResetButtonWrapper,
 	Header,
 	ImageGallery,
 	Navigation,
