+++ code/recipe-book-app/src/features/Appearance/AppearanceModificationForm/index.js
@@ -1,0 +1,2 @@
+import { AppearanceModificationForm } from "./AppearanceModificationForm";
+export default AppearanceModificationForm;
